move sidebar component files to separate folder
