+++ src/components/sidebar/index.js
@@ -1,0 +1,3 @@
+export { default as Sidebar } from './sidebar';
+export { default as SidebarDetails } from './details';
+export { default as SidebarOptions } from './options';
