+++ src/components/sidebar/options.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Button } from 'primereact/button';
 import { DataView } from 'primereact/dataview';
-import { updateItemAction } from '../redux/actions/action-dispatchers';
+import { updateItemAction } from '../../redux/actions/action-dispatchers';
 
 const SidebarOptions = ({ dataOptions }) => {
     const dispatch = useDispatch();
@@ -44,7 +44,6 @@ const SidebarOptions = ({ dataOptions }) => {
                     <img src={imgPath} alt={resultName} />
                     <div className='product-list-detail'>
                         <h5 className='p-mb-2'>{resultName}</h5>
-                        {/* <span className='product-list-badge'>{thing.status}</span> */}
                         <h6 className='p-mt-0 p-mb-2'>{JSON.stringify(data)}</h6>
                     </div>
                     <div className='product-list-action'>
